+++ include/wfa/signal_handler.hpp
@@ -1,0 +1,10 @@
+#ifndef WFA_SIGNAL_HANDLER_HPP
+#define WFA_SIGNAL_HANDLER_HPP
+
+namespace wfa {
+
+void InstallSignalHandler();
+
+}  // namespace wfa
+
+#endif  // WFA_SIGNAL_HANDLER_HPP